update logic controller for feat: admin
