+++ src/main/java/vn/edu/hcmaf/apigamestore/product/controller/adminController/AdminAccountController.java
@@ -1,0 +1,11 @@
+package vn.edu.hcmaf.apigamestore.product.controller.adminController;
+
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/admin/accounts")
+@Validated
+public class AdminAccountController {
+}
