+++ src/main/java/vn/edu/hcmaf/apigamestore/user/controller/userController/UserController.java
@@ -1,4 +1,4 @@
-package vn.edu.hcmaf.apigamestore.user;
+package vn.edu.hcmaf.apigamestore.user.controller.userController;
 
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
@@ -16,6 +16,9 @@ import vn.edu.hcmaf.apigamestore.common.dto.LazyLoadingRequestDto;
 import vn.edu.hcmaf.apigamestore.common.dto.LazyLoadingResponseDto;
 import vn.edu.hcmaf.apigamestore.common.dto.SuccessResponse;
 import vn.edu.hcmaf.apigamestore.product.dto.AccountDto;
+import vn.edu.hcmaf.apigamestore.user.UserEntity;
+import vn.edu.hcmaf.apigamestore.user.UserRepository;
+import vn.edu.hcmaf.apigamestore.user.UserService;
 import vn.edu.hcmaf.apigamestore.user.dto.UpdateUserDto;
 
 import javax.swing.text.html.parser.Entity;
@@ -28,38 +31,35 @@ import java.util.List;
 public class UserController {
     @Autowired
     private UserService userService;
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private CartService cartService;
+
+//    @GetMapping("/all")
+//    @PreAuthorize("hasAuthority('ADMIN')")
+//    public ResponseEntity<BaseResponse> getAllUsers() {
+//        List<UserEntity> users = userService.getAllUsers();
+//        return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get all User info success", users));
+//    }
+//    @PostMapping("/all-lazyloading")
+//    @PreAuthorize("hasAuthority('ADMIN')")
+//    public ResponseEntity<BaseResponse> getAllUsersLazyLoading(@RequestBody @Valid LazyLoadingRequestDto<Sort> lazyLoadingRequestDto) {
+//      LazyLoadingResponseDto<List<UserEntity>> users = userService.getAllUsersLazyLoading(lazyLoadingRequestDto);
+//      return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get all User info success (lazyloading)", users));
+//    }
+//    @GetMapping("/id/{userId}")
+//    @PreAuthorize("hasAuthority('ADMIN')")
+//    public ResponseEntity<BaseResponse> getUserById(@PathVariable Long userId) {
+//      UserEntity userEntity = userService.getUserById(userId);
+//      return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get User Id : "+ userId+" info success", userEntity));
+//    }
+//    @GetMapping("/username/{username}")
+//    @PreAuthorize("hasAuthority('ADMIN')")
+//    public ResponseEntity<BaseResponse> getUserByUsername(@PathVariable String username) {
+//      UserEntity userEntity = userService.getUserByEmail(username);
+//      return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get User Name : "+ username+" info success", userEntity));
+//    }
     @GetMapping("/me")
     public ResponseEntity<BaseResponse> getCurrentUser() {
         UserEntity userEntity = userService.getCurrentUser();
         return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get User info success", userEntity));
-    }
-    @GetMapping("/all")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<BaseResponse> getAllUsers() {
-        List<UserEntity> users = userService.getAllUsers();
-        return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get all User info success", users));
-    }
-    @PostMapping("/all-lazyloading")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<BaseResponse> getAllUsersLazyLoading(@RequestBody @Valid LazyLoadingRequestDto<Sort> lazyLoadingRequestDto) {
-      LazyLoadingResponseDto<List<UserEntity>> users = userService.getAllUsersLazyLoading(lazyLoadingRequestDto);
-      return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get all User info success (lazyloading)", users));
-    }
-    @GetMapping("/id/{userId}")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<BaseResponse> getUserById(@PathVariable Long userId) {
-      UserEntity userEntity = userService.getUserById(userId);
-      return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get User Id : "+ userId+" info success", userEntity));
-    }
-    @GetMapping("/username/{username}")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<BaseResponse> getUserByUsername(@PathVariable String username) {
-      UserEntity userEntity = userService.getUserByEmail(username);
-      return ResponseEntity.ok().body(new SuccessResponse<>("SUCCESS","Get User Name : "+ username+" info success", userEntity));
     }
     @PutMapping("/{userId}")
     public ResponseEntity<BaseResponse> updateUser(@RequestBody @Valid  UpdateUserDto updateUserDto, @PathVariable Long userId) {
